refactor: relocate command (API)
